+++ plugins/relay/src/main/java/com/anezium/rokidbus/plugin/relay/RelayPrivacy.kt
@@ -1,0 +1,6 @@
+package com.anezium.rokidbus.plugin.relay
+
+/** What the glasses show in place of a message body while a hide switch is on. */
+internal object RelayPrivacy {
+    const val HIDDEN_BODY = "New message"
+}